% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_dir}
\alias{copy_dir}
\title{Copy directory}
\usage{
copy_dir(src, dst)
}
\arguments{
\item{src}{Character. Source directory.}

\item{dst}{Character. Destination directory.}
}
\value{
Invisibly returns NULL.
}
\description{
Copies contents of a source directory to a destination directory.
}
\keyword{internal}
