% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best.R
\name{get_best}
\alias{get_best}
\title{Get best optimization record}
\usage{
get_best(obj_txt)
}
\arguments{
\item{obj_txt}{Character. Path to objective log file.}
}
\value{
Data frame containing the best record.
}
\description{
Reads the objective log file and returns the best record.
}
