% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatia.R
\name{get_best_obj}
\alias{get_best_obj}
\title{Get best objective value}
\usage{
get_best_obj(obj_day_txt)
}
\arguments{
\item{obj_day_txt}{Character. Path to objective log file.}
}
\value{
Numeric scalar.
}
\description{
Returns the minimum objective value from the log file.
}
