% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatia.R
\name{get_best_x}
\alias{get_best_x}
\title{Get best normalized parameter vector}
\usage{
get_best_x(obj_day_txt)
}
\arguments{
\item{obj_day_txt}{Character. Path to objective log file.}
}
\value{
Numeric vector.
}
\description{
Returns the normalized parameter vector corresponding to the best
objective value.
}
