% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_swatia_cli}
\alias{install_swatia_cli}
\title{Install swatia CLI into PATH}
\usage{
install_swatia_cli(dir = file.path(Sys.getenv("HOME"), ".local", "bin"))
}
\arguments{
\item{dir}{Character. Directory in which to create the \code{swatia}
symlink. This directory should be included in the user's \code{PATH}.}
}
\description{
Creates a symbolic link named \code{swatia} in the specified directory so
that the SWATIA CLI can be invoked from the command line. The target
directory must be included in the user's \code{PATH}.
}
