% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatia.R
\name{print_list}
\alias{print_list}
\title{Print named list}
\usage{
print_list(x)
}
\arguments{
\item{x}{List.}
}
\value{
Invisibly returns NULL.
}
\description{
Prints elements of a named list in a readable format.
}
\keyword{internal}
