% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_dir}
\alias{reset_dir}
\title{Reset directories}
\usage{
reset_dir(paths)
}
\arguments{
\item{paths}{Character vector. Directory paths.}
}
\value{
Invisibly returns NULL.
}
\description{
Removes and recreates specified directories.
}
\keyword{internal}
