% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatia.R
\name{run_swatplus_in_dir}
\alias{run_swatplus_in_dir}
\title{Run SWAT+ in a directory}
\usage{
run_swatplus_in_dir(dir)
}
\arguments{
\item{dir}{Character. Path to a SWAT+ TxtInOut_* directory.}
}
\value{
Exit status from \code{system2()}.
}
\description{
Executes the SWAT+ model in the specified directory.
}
\keyword{internal}
