% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{extract_sim}
\alias{extract_sim}
\title{Extract simulated discharge}
\usage{
extract_sim(dir, chaid)
}
\arguments{
\item{dir}{Character. Path to a SWAT+ TxtInOut_* directory.}

\item{chaid}{Integer or character. Channel ID.}
}
\value{
Numeric vector of simulated discharge values.
}
\description{
Reads simulated discharge values for a given channel ID from a SWAT+ output
directory.
}
\keyword{internal}
